requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
dj-rest-auth
allauth
PyYAML==6.0.1
drf-yasg==1.21.7